feat(scripts): refactor `search_versions` to generic search framework
