+++ scripts/codesearch/README.md
@@ -1,0 +1,7 @@
+# Codesearch
+
+This folder contains scripts that help to find all kind of patterns in the codebase.
+
+## search_versions
+
+Helps to find all occurrences of version numbers in the rust code (e.g. function names, variable names, etc.).
