--- scripts/search_versions/README.md
@@ -1,3 +1,0 @@
-# Search versions
-
-This script helps to find all occurrences of version numbers in the rust code (e.g. function names, variable names, etc.).
